Correcting AI's wrong codes
Correct AI's codes by comparing with the original main branch
swin_transformer.py's modification is incomplete

TODO:
- groundingdino/models/GroundingDINO/backbone/swin_transformer.py
- groundingdino/models/GroundingDINO/bertwraper.py
- groundingdino/models/GroundingDINO/fuse_modules.py
- groundingdino/models/GroundingDINO/groundingdino.py
- groundingdino/models/GroundingDINO/ms_deform_attn.py
- groundingdino/models/GroundingDINO/transformer.py
- groundingdino/models/GroundingDINO/transformer_vanilla.py
- groundingdino/models/GroundingDINO/utils.py
- groundingdino/models/GroundingDINO/ops/jittor_ms_deform_attn.py
- groundingdino/util/box_ops.py
- groundingdino/util/debug_tools.py
- groundingdino/util/get_tokenlizer.py
- groundingdino/util/inference.py
- groundingdino/util/misc.py
- groundingdino/util/utils.py
- groundingdino/util/visualizer.py
- groundingdino/util/vl_utils.py

+++ GroundingDINO/groundingdino/models/GroundingDINO/ops/jittor_ms_deform_attn.py
@@ -17,7 +17,7 @@ def ms_deform_attn_core_jittor(value, value_spatial_shapes, sampling_locations, 
     """
     # convert shapes to python list if jt.Var
     if isinstance(value_spatial_shapes, jt.Var):
-        vsh = value_spatial_shapes.numpy().astype(int).tolist()
+        vsh = value_spatial_shapes.numpy().int().tolist()
     else:
         vsh = [tuple(map(int, x)) for x in value_spatial_shapes]
